+++ backend/src/main/java/dev/tripdraw/post/application/PostQueryService.java
@@ -1,0 +1,4 @@
+package dev.tripdraw.post.application;
+
+public class PostQueryService {
+}
