--- backend/src/main/java/dev/tripdraw/post/application/query/PostQueryService.java
@@ -1,5 +1,0 @@
-package dev.tripdraw.post.application.query;
-
-public class PostQueryService {
-
-}
