+++ backend/src/main/java/dev/tripdraw/post/domain/PostRepository.java
@@ -2,9 +2,8 @@ package dev.tripdraw.post.domain;
 
 import static dev.tripdraw.post.exception.PostExceptionType.POST_NOT_FOUND;
 
-import dev.tripdraw.post.domain.query.PostCustomRepository;
 import dev.tripdraw.post.exception.PostException;
-import lombok.NonNull;
+import dev.tripdraw.post.query.PostCustomRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -13,8 +12,7 @@ public interface PostRepository extends JpaRepository<Post, Long>, PostCustomRep
 
     List<Post> findAllByTripId(Long tripId);
 
-    @NonNull
-    default Post getById(@NonNull Long id) {
+    default Post getByPostId(Long id) {
         return findById(id)
                 .orElseThrow(() -> new PostException(POST_NOT_FOUND));
     }
