+++ backend/src/main/java/dev/tripdraw/post/dto/PostSearchPaging.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.post.dto.query;
+package dev.tripdraw.post.dto;
 
 public record PostSearchPaging(Long lastViewedId, Integer limit) {
 }
