+++ backend/src/main/java/dev/tripdraw/post/query/PostCustomRepository.java
@@ -1,8 +1,8 @@
-package dev.tripdraw.post.domain.query;
+package dev.tripdraw.post.query;
 
 import dev.tripdraw.post.domain.Post;
-import dev.tripdraw.post.dto.query.PostSearchConditions;
-import dev.tripdraw.post.dto.query.PostSearchPaging;
+import dev.tripdraw.post.dto.PostSearchConditions;
+import dev.tripdraw.post.dto.PostSearchPaging;
 
 import java.util.List;
