+++ backend/src/test/java/dev/tripdraw/draw/application/PostCreateEventHandlerIntegrationTest.java
@@ -22,7 +22,7 @@ import org.springframework.transaction.support.TransactionTemplate;
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SpringBootTest
-public class PostCreateEventHandlerIntegrationTest {
+class PostCreateEventHandlerIntegrationTest {
 
     @MockBean
     private RouteImageGenerator routeImageGenerator;
@@ -45,7 +45,7 @@ public class PostCreateEventHandlerIntegrationTest {
         PostCreateEvent postCreateEvent = new PostCreateEvent(1L, 1L);
         given(tripRepository.getTripWithPoints(postCreateEvent.tripId()))
                 .willReturn(여행());
-        given(postRepository.getById(postCreateEvent.postId()))
+        given(postRepository.getByPostId(postCreateEvent.postId()))
                 .willReturn(감상());
 
         // when
