+++ backend/src/test/java/dev/tripdraw/post/application/PostServiceTest.java
@@ -25,13 +25,13 @@ import dev.tripdraw.post.dto.PostAndPointCreateRequest;
 import dev.tripdraw.post.dto.PostCreateResponse;
 import dev.tripdraw.post.dto.PostRequest;
 import dev.tripdraw.post.dto.PostResponse;
+import dev.tripdraw.post.dto.PostSearchConditions;
+import dev.tripdraw.post.dto.PostSearchPaging;
 import dev.tripdraw.post.dto.PostSearchRequest;
 import dev.tripdraw.post.dto.PostSearchResponse;
 import dev.tripdraw.post.dto.PostUpdateRequest;
 import dev.tripdraw.post.dto.PostsResponse;
 import dev.tripdraw.post.dto.PostsSearchResponse;
-import dev.tripdraw.post.dto.query.PostSearchConditions;
-import dev.tripdraw.post.dto.query.PostSearchPaging;
 import dev.tripdraw.post.exception.PostException;
 import dev.tripdraw.test.ServiceTest;
 import dev.tripdraw.trip.domain.Point;
