+++ backend/src/test/java/dev/tripdraw/post/domain/PostCustomRepositoryImplTest.java
@@ -7,9 +7,9 @@ import dev.tripdraw.common.config.JpaConfig;
 import dev.tripdraw.common.config.QueryDslConfig;
 import dev.tripdraw.member.domain.Member;
 import dev.tripdraw.member.domain.MemberRepository;
-import dev.tripdraw.post.domain.query.PostCustomRepositoryImpl;
-import dev.tripdraw.post.dto.query.PostSearchConditions;
-import dev.tripdraw.post.dto.query.PostSearchPaging;
+import dev.tripdraw.post.dto.PostSearchConditions;
+import dev.tripdraw.post.dto.PostSearchPaging;
+import dev.tripdraw.post.query.PostCustomRepositoryImpl;
 import dev.tripdraw.trip.domain.Point;
 import dev.tripdraw.trip.domain.Trip;
 import dev.tripdraw.trip.domain.TripName;
